Add Touch Command
modify Mkdir command
add CreateEntry command
add Touch command
